+++ src/components/button/Button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
-import Button, { ButtonProps } from './Button';
+import Button, { ButtonProps } from './index';
 
 export default {
   title: 'Button/Button',
