+++ src/components/form/Checkbox/Checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
-import Checkbox, { CheckboxProps } from './Checkbox';
+import Checkbox, { CheckboxProps } from './index';
 
 export default {
   title: 'Form/Checkbox',
