+++ src/components/form/InputText/InputText.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, StoryFn } from '@storybook/react';
-import InputText, { InputTextProps } from './InputText';
+import InputText, { InputTextProps } from './index';
 
 export default {
   title: 'Form/InputText',
