+++ src/components/messages/Toast/index.tsx
@@ -1,4 +1,5 @@
 import React, { createContext, useContext, useState, ReactNode } from 'react';
+import './Toast.css';
 import { createPortal } from 'react-dom';
 
 type ToastPosition = 'top-left' | 'top-right' | 'top-center' | 'bottom-left' | 'bottom-right' | 'bottom-center';
