+++ src/components/panel/ScrollPanel/ScrollPanel.css
@@ -1,0 +1,34 @@
+.scrollpanel {
+  overflow: hidden;
+  position: relative;
+  height: 100%;
+}
+
+.scrollpanel-content {
+  overflow: auto;
+  height: 100%;
+  scrollbar-width: var(--scrollbar-width, 12px);
+}
+
+.scrollpanel-content::-webkit-scrollbar {
+  width: var(--scrollbar-width, 12px);
+  height: var(--scrollbar-width, 12px);
+}
+
+.scrollpanel-content::-webkit-scrollbar-track {
+  background: var(--scrollbar-bgcolor, #f5f5f5);
+}
+
+.scrollpanel-content::-webkit-scrollbar-thumb {
+  background-color: var(--scrollbar-color, #555);
+  border-radius: 10px;
+  border: 2px solid var(--scrollbar-bgcolor, #f5f5f5);
+}
+
+.scrollpanel-content::-webkit-scrollbar-thumb:hover {
+  background-color: var(--scrollbar-color, #444);
+}
+
+.scrollpanel-content::-webkit-scrollbar-corner {
+  background: transparent;
+}
