+++ src/index.ts
@@ -1,8 +1,8 @@
-import Button from './components/Button';
-import Skeleton from './components/Skeleton';
-import Checkbox from './components/Checkbox';
-import InputText from './components/InputText';
-import ScrollPanel from './components/ScrollPanel';
-import { ToastProvider, useToast } from './components/Toast';
+import Button from './components/button/Button';
+import Skeleton from './components/misc/Skeleton';
+import Checkbox from './components/form/Checkbox';
+import InputText from './components/form/InputText';
+import ScrollPanel from './components/panel/ScrollPanel';
+import { ToastProvider, useToast } from './components/messages/Toast';
 
 export { Button, Skeleton, Checkbox, InputText, ScrollPanel, ToastProvider, useToast };
